backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
aiofiles
websockets
gTTS
